fix(aws/cloudformation): Disable cloudformation cache agent
